Change sleep location and switch to pipe for IPC
Use a simple multiprocessing pipe to delay exiting the parent process
until after the child has been doubly forked and has had a chance to
reach the code that will read the module to be executed asynchronously.

This reduces a race on heavily loaded systems, where async_wrapper can
return and the calling code will continue onwards and delete the module
files from the remote before the grand-child process has a chance to
being reading the module code to be executed asynchronously.

In such cases the following error will be seen:

 msg: [Errno 2] No such file or directory

The use of some simple IPC and changing the sleep to occur after the
subprocess call ensures that the started process should be finished
reading the given files before telling the waiting grandparent process
that it's ok to return.